hoping intermittent error not coming back
