+++ lib/board_selections.rb
@@ -1,5 +1,5 @@
 module BoardSelections
-  
+
   def board
     ["A1","A2","A3","A4",
      "B1","B2","B3","B4",
@@ -26,4 +26,13 @@ module BoardSelections
       ["C4","C3"]=>["C2"], ["C4","D4"]=>["B4"], ["D1","C1"]=>["B1"], ["D1","D2"]=>["D3"], ["D2","D1"]=>["D3"], ["D2","C2"]=>["B2"],
       ["D3","C3"]=>["B3"], ["D3","D2"]=>["D1","D4"], ["D3","D4"]=>["D2"], ["D4","C4"]=>["B4"], ["D4","D3"]=>["D2"]}
   end
+
+  # def two_unit_sunk?
+  #   return true if two_unit_ship == 2
+  # end
+  #
+  # def three_unit_sunk?
+  #   return true if three_unit_ship == 3
+  # end
+
 end
