+++ lib/messages.rb
@@ -38,6 +38,23 @@ module Messages
     `say -r 180 "Just what do you think you're doing, Dave? Place the ship within the defined coordinates on the board"`
   end
 
+  def game_over
+    if computer.two_unit_sunk? && computer.three_unit_sunk?
+      exit_message
+      puts "(Y)es or (N)o?"
+      input = gets.chomp
+      if yes_commands(input)
+        song
+      else
+        p "Goodbye!"
+        exit
+      end
+    elsif player.two_unit_sunk? && player.three_unit_sunk?
+      you_lose
+      exit
+    end
+  end
+
   def you_lose
         banner = "
                   ┼┼┼┼┼┼┼┼┼┼┼┼┼┼┼┼┼┼┼┼┼┼┼┼┼┼┼┼┼┼┼┼┼┼┼┼
